requirements.txt
uiprotect>=7.0.0
asyncio
python-dotenv>=1.0.0
colorama>=0.4.0
click>=8.0.0
RPi.GPIO>=0.7.0 